Adding bittorrent protocol example.
This library was originally built for a bittorrent implementation.  The wire
protocol for that implementation is the bittorrent example here.  I've removed
it from the actual library, which deserves to be cleaned up and released on
it's own when I get the time.  I think this is a good example of how
ProtoStruct is intended to work: define the wire protocol, mix it into your
socket, and get parsed messages.

I also fixed some typos in the comments of lib/buffered_stream.rb.